Republishing packages as scoped packages

+++ republish-npm-packages.sh
@@ -27,7 +27,7 @@ EOT
 if ! [[ -n "$CURRENT_PACKAGE" ]]; then show_usage; fi;
 if ! [[ -n "$NEW_PACKAGE_NAME" ]]; then show_usage; fi;
 
-if ! command -v foo 2>/dev/null; then
+if ! command -v jq 2>/dev/null; then
     echo ""
     echo -e "\033[0;31mJQ is required for package.json manipulation.\033[0m"
     echo "Download it at https://stedolan.github.io/jq/ and make sure to have jq available in your path."
@@ -35,53 +35,59 @@ if ! command -v foo 2>/dev/null; then
     exit 1
 fi
 
-exit 0
-
 mkdir -p packages/$CURRENT_PACKAGE
-CWD=`pwd`/$CURRENT_PACKAGE
+PWD=`pwd`
+CWD="${PWD}/packages/$CURRENT_PACKAGE"
+
 cd $CWD
+echo $CWD
 
 # Fetching and downloading all versions for specified package
 VERSIONS=`npm view $CURRENT_PACKAGE versions | grep -E -o "\d+\.\d+\.\d+"`
 echo "Versions returned for $CURRENT_PACKAGE: $VERSIONS "
 
+cd $CWD
 
 for VERSION in $VERSIONS
 do
-    cd $CWD
-
-    # Downloads the tarball with the specified version
-    npm pack $CURRENT_PACKAGE@$VERSION
-
-    # Create a directory for each version, to receive the unpacked package
-    mkdir $VERSION
-
-    # Unpacks files to the corresponding directory.
-    #   strip-components=1 removes the /package directory from the tarball
-    #   -C defines the destination path
-    tar -xvzf $CURRENT_PACKAGE-$VERSION.tgz --strip-components=1 -C $VERSION
-
-    cd $VERSION
-
-    # Changing package name
-    cat package.json | jq --arg NAME "$NEW_PACKAGE_NAME" '.name = $NAME' > package.json
-
-    # Displaying new name saved on package.json
-    cat package.json | jq '{.name}'
-
-    rm ../$CURRENT_PACKAGE-$VERSION.tgz
-    cd $CWD
+    # Grabs the tarballs from npm registry with the specified version
+    # Downloads are made in parallel
+    npm pack $CURRENT_PACKAGE@$VERSION &
 done
 
+wait
 
-for $VERSION in $VERSIONS
-do
-    cd $CWD/$VERSION
+if [[ $PUBLISH != '--publish' ]]; then
+    for VERSION in $VERSIONS
+    do
+        cd $CWD
 
-    # We need to skip the execution of npm publish scripts.
-    # So we'll pack them to tgz first, then upload.
-    npm pack
-    npm publish --access=public
+        # Create a directory for each version, to receive the unpacked package
+        mkdir $VERSION
 
+        # Unpacks files to the corresponding directory.
+        #   strip-components=1 removes the /package directory from the tarball
+        #   -C defines the destination path
+        tar -xvzf $CURRENT_PACKAGE-$VERSION.tgz --strip-components=1 -C $VERSION
 
-done
+        cd $VERSION
+        pwd
+
+        # # Changing package name
+        cat package.json | jq --arg NAME "$NEW_PACKAGE_NAME" '.name = $NAME' > package.json1
+        mv package.json1 package.json
+
+        # # Displaying new name saved on package.json
+        cat package.json | jq '.name'
+
+        rm ../$CURRENT_PACKAGE-$VERSION.tgz
+
+        # We need to skip the execution of npm publish scripts, 
+        # so we'll pack them to tgz first, then upload.
+        npm pack
+
+        mv *$VERSION.tgz ../
+    done
+fi
+
+echo "DONE"
